backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
python-dotenv>=0.19.0

psycopg2-binary>=2.9.1

python-dotenv>=0.19.0